// WADNR.Web/src/app/shared/generated/model/focus-area-grid-row.ts
export class FocusAreaGridRow { 
    FocusAreaID?: number;
    FocusAreaName?: string | null;
    FocusAreaStatusID?: number;
    FocusAreaStatusDisplayName?: string | null;
    DNRUplandRegionID?: number;
    DNRUplandRegionName?: string | null;
    PlannedFootprintAcres?: number | null;
    ProjectCount?: number;
    HasLocation?: boolean;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}